# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_web_controllers_http.SwaggerCategoryListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Category'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerCategoryResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Category'
    type: object
...
  internal_web_controllers_http.SwaggerCreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  internal_web_controllers_http.SwaggerCreateProductRequest:
    properties:
      basePrice:
        type: number
      categoryId:
        type: string
      commissionType:
        type: string
      commissionValue:
        type: number
      cost:
        type: number
      defaultLeadDays:
        type: integer
      name:
        type: string
      notes:
        type: string
      requiresDelivery:
        type: boolean
      type:
        type: string
    required:
    - categoryId
    - name
    - type
    type: object
...
  internal_web_controllers_http.SwaggerErrorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          type: string
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  internal_web_controllers_http.SwaggerMeta:
    properties:
      total:
        type: integer
    type: object
  internal_web_controllers_http.SwaggerOKFlag:
    properties:
      ok:
        type: boolean
    type: object
  internal_web_controllers_http.SwaggerOKFlagResponse:
    properties:
      data:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlag'
    type: object
...
  internal_web_controllers_http.SwaggerProductListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Product'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerProductResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.Product'
    type: object
...
  internal_web_controllers_http.SwaggerUpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  internal_web_controllers_http.SwaggerUpdateProductRequest:
    properties:
      basePrice:
        type: number
      categoryId:
        type: string
      commissionType:
        type: string
      commissionValue:
        type: number
      cost:
        type: number
      defaultLeadDays:
        type: integer
      name:
        type: string
      notes:
        type: string
      requiresDelivery:
        type: boolean
      type:
        type: string
    required:
    - categoryId
    - name
    - type
    type: object
...
  photogallery_api_go_internal_domain_entities.Category:
    properties:
      createdAtUtc:
        type: string
      description:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      updatedAtUtc:
        type: string
    type: object
...
  photogallery_api_go_internal_domain_entities.Product:
    properties:
      basePrice:
        format: float64
        type: number
      categoryID:
        type: string
      commissionType:
        type: string
      commissionValue:
        format: float64
        type: number
      cost:
        format: float64
        type: number
      createdAtUtc:
        type: string
      defaultLeadDays:
        type: integer
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      notes:
        type: string
      requiresDelivery:
        type: boolean
      type:
        type: string
      updatedAtUtc:
        type: string
    type: object
...
info:
  contact: {}
  description: API REST v1 Foto-tienda
  title: Photo Gallery API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      description: Lista paginada de categorías con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCategoryListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListCategories
      tags:
      - catalog
    post:
      consumes:
      - application/json
      description: Crea una nueva categoría.
      parameters:
      - description: Create category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateCategory
      tags:
      - catalog
  /api/v1/categories/:id:
    get:
      description: Obtiene una categoría por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetCategory
      tags:
      - catalog
    put:
      consumes:
      - application/json
      description: Actualiza una categoría existente.
      parameters:
      - description: Update category payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateCategoryRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateCategory
      tags:
      - catalog
  /api/v1/categories/:id/activate:
    patch:
      description: Activa una categoría.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ActivateCategory
      tags:
      - catalog
  /api/v1/categories/:id/deactivate:
    patch:
      description: Desactiva una categoría.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeactivateCategory
      tags:
      - catalog
...
  /api/v1/products:
    get:
      description: Lista paginada de productos con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerProductListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListProducts
      tags:
      - catalog
    post:
      consumes:
      - application/json
      description: Crea un nuevo producto.
      parameters:
      - description: Create product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateProduct
      tags:
      - catalog
  /api/v1/products/:id:
    get:
      description: Obtiene un producto por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetProduct
      tags:
      - catalog
    put:
      consumes:
      - application/json
      description: Actualiza un producto existente.
      parameters:
      - description: Update product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateProductRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateProduct
      tags:
      - catalog
  /api/v1/products/:id/activate:
    patch:
      description: Activa un producto.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ActivateProduct
      tags:
      - catalog
  /api/v1/products/:id/deactivate:
    patch:
      description: Desactiva un producto.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeactivateProduct
      tags:
      - catalog
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"